refactor tests and remove extraneous files
