--- src/test.ts
@@ -1,83 +1,0 @@
-// var hydration = require('./lib/NestHydrationJS')();
-// let hydrate = hydration();
-
-// const data = [{
-//     'group_has_service.service_id': 2,
-//     'group_has_service.group_id': 5,
-//     'zone.id': 5,
-//     'zone.name': 'Wings',
-//     'service.id': 2,
-//     'service.name': 'Deep Clean',
-//     'group.id': 5,
-//     'group.name': 'Group 5!'
-// },
-// {
-//     'group_has_service.service_id': 2,
-//     'group_has_service.group_id': 5,
-//     'zone.id': 4,
-//     'zone.name': 'Struts',
-//     'service.id': 2,
-//     'service.name': 'Deep Clean',
-//     'group.id': 5,
-//     'group.name': 'Group 5!'
-// },
-// {
-//     'group_has_service.service_id': 2,
-//     'group_has_service.group_id': 5,
-//     'zone.id': 3,
-//     'zone.name': 'Wheels',
-//     'service.id': 2,
-//     'service.name': 'Deep Clean',
-//     'group.id': 5,
-//     'group.name': 'Group 5!'
-// },
-// {
-//     'group_has_service.service_id': 3,
-//     'group_has_service.group_id': 5,
-//     'zone.id': 3,
-//     'zone.name': 'Wheels',
-//     'service.id': 3,
-//     'service.name': 'Quick Wash',
-//     'group.id': 5,
-//     'group.name': 'Group 5!'
-// },
-// {
-//     'group_has_service.service_id': 3,
-//     'group_has_service.group_id': 5,
-//     'zone.id': 1,
-//     'zone.name': 'Elevators',
-//     'service.id': 3,
-//     'service.name': 'Quick Wash',
-//     'group.id': 5,
-//     'group.name': 'Group 5!'
-// },
-// {
-//     'group_has_service.service_id': 3,
-//     'group_has_service.group_id': 5,
-//     'zone.id': 6,
-//     'zone.name': 'APU',
-//     'service.id': 3,
-//     'service.name': 'Quick Wash',
-//     'group.id': 5,
-//     'group.name': 'Group 5!'
-// },]
-
-// const description = [{
-//     service_id: {column: 'group_has_service.service_id', id: true},
-//     group_id: {column: 'group_has_service.group_id'},
-//     service: {
-//         id: 'service.id',
-//         name: 'service.name',
-//         zone: [{
-//             id: 'zone.id',
-//             name: 'zone.name'
-//         }]
-//     },
-//     group: {
-//         id: 'group.id',
-//         name: 'group.name'
-//     }
-// }]
-
-// const res = hydrate.nest(data, description);
-// console.log(res);
